Delphi grpc service.  Generic api server. segregate service protos, services
